Add movement and correct colliders
Colliders now change to the correct collider for the given shape. You can now move left and right with the circle however there is a bug where angular velocity/angular acceleration (torque) is not reduced to zero upon stopping after changing to the square or triangle. This means that sometimes when you move with the circle, change into a triangle and stop, and change back, you'll start moving again without pressing any buttons.